[BUILD] Update to Intellij 2018.2.6
This updated is needed to build saros-project/saros#343. The problem is
that the new logic relies on the fact that IntelliJ listeners now have
default implementations. In the currently used version, the default
implementation for the methods of FileEditorManagerListener.Before()
does not seem to exist yet. This causes the build to fail.

Even though these compatibility issues are resolved with 2018.2,
2018.2.6 is still used as it is the earliest version of the major
release 2018.2.* that is still publicly available.